+++ bms/main.py
@@ -17,6 +17,8 @@ import configparser
 
 sys.path.append('.')
 
+from bms.v1.listeners import EventListener
+
 define("port", default=8888, help="Tornado Web port", type=int)
 define("pg_user", default="postgres", help="PostgrSQL database user")
 define("pg_password", default="postgres", help="PostgrSQL user password")
@@ -25,16 +27,67 @@ define("pg_port", default="5432", help="PostgrSQL database port")
 define("pg_database", default="bms", help="PostgrSQL database name")
 define("pg_upgrade", default=False, help="Upgrade PostgrSQL schema", type=bool)
 
-define("file_repo", default='local', help="Select the file repository", type=str)
+define("file_repo", default='s3', help="Select the file repository", type=str)
 
 # Local storage for files configuration
 define("local_path", default=str(Path.home()), help="Select local path", type=str)
 
-# AWS S3 stzorage for files configuration
+# AWS S3 storage for files configuration
 define("aws_bucket", default='bdms', help="Select AWS Bucket name", type=str)
 define("aws_credentials", default=None, help="AWS S3 credential file location (overwrite aws_access_key_id and aws_secret_access_key)", type=str)
 define("aws_access_key_id", default=None, help="AWS S3 access key id", type=str)
 define("aws_secret_access_key", default=None, help="AWS S3 secret access key", type=str)
+
+# SMTP send mail configuration
+define(
+    "smtp_config",
+    default=None,
+    help="SMTP configuration file location",
+    type=str
+)
+define(
+    "smtp_recipients",
+    default=None,
+    help="SMTP comma separated recipients email addresses",
+    type=str
+)
+define(
+    "smtp_username",
+    default=None,
+    help="SMTP username",
+    type=str
+)
+define(
+    "smtp_password",
+    default=None,
+    help="SMTP password",
+    type=str
+)
+define(
+    "smtp_server",
+    default=None,
+    help="SMTP server address",
+    type=str
+)
+define(
+    "smtp_port",
+    default=587,
+    help="SMTP server port",
+    type=int
+)
+define(
+    "smtp_tls",
+    default=False,
+    help="SMTP server supports direct connection via TLS/SSL",
+    type=bool
+)
+define(
+    "smtp_starttls",
+    default=True,
+    help="SMTP servers support the STARTTLS extension",
+    type=bool
+)
+
 
 # Ordered list of available versions
 versions = [
@@ -50,6 +103,8 @@ sql_files = {
     "1.0.0": f"{udir}1.0.0_to_1.0.1-RC1.sql"
 }
 
+listeners = []
+
 async def get_conn():
     try:
         return await asyncpg.create_pool(
@@ -62,10 +117,6 @@ async def get_conn():
     except Exception as x:
         red("Connection to db failed.")
         raise x
-
-
-async def release_pool(pool):
-    await pool.close()
 
 def red(message):
     print(f"\033[91m{message}\033[0m")
@@ -158,6 +209,19 @@ async def system_check(pool):
             current_db_version
         )
 
+async def close(application):
+    # Remove all listeners
+    blue("Removing listeners..")
+    await application.listener.stop()
+    green(" > done.")
+
+    blue("Closing connection pool..")
+    # Closing connections pool
+    await application.pool.close()
+    green(" > done.")
+
+    print("")
+
 if __name__ == "__main__":
 
     options.parse_command_line()
@@ -202,6 +266,9 @@ if __name__ == "__main__":
         # Terms handlers
         TermsHandler,
         TermsAdminHandler,
+
+        # Feedback handler
+        FeedbackHandler,
 
         # Other handlers
         GeoapiHandler,
@@ -251,6 +318,9 @@ if __name__ == "__main__":
         (r'/api/v1/terms', TermsHandler),
         (r'/api/v1/terms/admin', TermsAdminHandler),
 
+        # FEEDBACK handlers
+        (r'/api/v1/feedback', FeedbackHandler),
+
         # Stratigraphy handlers
         (r'/api/v1/borehole/stratigraphy', StratigraphyViewerHandler),
         (r'/api/v1/borehole/stratigraphy/edit', StratigraphyProducerHandler),
@@ -271,14 +341,22 @@ if __name__ == "__main__":
 
     ], **settings)
 
+    # Init database postgresql connection pool
     application.pool = ioloop.run_until_complete(get_conn())
 
+    # Create events listeners
+    application.listener = EventListener(application)
+    ioloop.run_until_complete(application.listener.start())
+
+    # Init config file parser
+    config = configparser.ConfigParser()
+
+    # Configuring S3 credentials
     if options.file_repo == 's3':
 
         if options.aws_credentials is None:
             raise Exception("AWS Credential file missing")
         
-        config = configparser.ConfigParser()
         config.read(options.aws_credentials)
 
         if (
@@ -290,6 +368,42 @@ if __name__ == "__main__":
 
         options.aws_access_key_id = config['default']['aws_access_key_id']
         options.aws_secret_access_key = config['default']['aws_secret_access_key']
+
+    # Configuring SMTP credentials
+    if options.smtp_config is not None:
+        
+        config.read(options.smtp_config)
+
+        if (
+            'SMTP' not in config or
+            'smtp_recipients' not in config['SMTP'] or
+            'smtp_username' not in config['SMTP'] or
+            'smtp_password' not in config['SMTP'] or
+            'smtp_server' not in config['SMTP']
+        ):
+            raise Exception("SMTP config file wrong")
+
+        options.smtp_recipients = config['SMTP']['smtp_recipients']
+        options.smtp_username = config['SMTP']['smtp_username']
+        options.smtp_password = config['SMTP']['smtp_password']
+        options.smtp_server = config['SMTP']['smtp_server']
+
+        if 'smtp_port' in config['SMTP']:
+            options.smtp_port = int(config['SMTP']['smtp_port'])
+
+        if 'smtp_tls' in config['SMTP']:
+            options.smtp_tls = (
+                True
+                if config['SMTP']['smtp_tls'] == '1'
+                else False
+            )
+
+        if 'smtp_starttls' in config['SMTP']:
+            options.smtp_starttls = (
+                True
+                if config['SMTP']['smtp_starttls'] == '1'
+                else False
+            )
 
     try:
         # Check system before startup
@@ -342,5 +456,8 @@ if __name__ == "__main__":
     except Exception as ex:
         print(f"Exception:\n{ex}")
 
+    except KeyboardInterrupt:
+        red("\nKeyboard interruption, probably: CTR+C\n")
+
     finally:
-        ioloop.run_until_complete(release_pool(application.pool))
+        ioloop.run_until_complete(close(application))