+++ bms/v1/terms/__init__.py
@@ -1,5 +1,4 @@
-
-# Actions
+# -*- coding: utf-8 -*-
 from bms.v1.terms.accept import AcceptTerms
 from bms.v1.terms.draft import DraftTerms
 from bms.v1.terms.get import GetTerms